Update Snakefile
added `gdown` command for downloading ICEberg, even though the wget command seems to work locally. The primary difference is in the number of lines in the file. Also added a `gdown` link for KEGG as we will be depreciation official support for it in the future and are moving away from downloading it officially in the pipeline.